Remade module logicytics
Made it so each class is a new package, so its easier to fix and debug, also _debug.py fixed bug it compares json config not ini
