+++ CODE/logicytics/__init__.py
@@ -1,7 +1,16 @@
-from logicytics.__lib_class import *
+from logicytics.Execute import *
+from logicytics.Get import *
+from logicytics.Logger import *
+from logicytics.Checks import *
+from logicytics.FileManagement import *
+from logicytics.Flag import *
 
-DEBUG: str
-VERSION: str
-CURRENT_FILES: list[str]
+
+Execute = Execute()
+Get = Get()
+Check = Check()
+FileManagement = FileManagement()
+Flag = Flag()
+
 
 DEBUG, VERSION, CURRENT_FILES, DELETE_LOGS = Get.config_data()
